# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/rides": {
            "post": {
                "description": "create ride",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rides"
                ],
                "summary": "starts a ride.",
                "parameters": [
                    {
                        "description": "Ride request parameters",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RideRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rides.Ride"
                        }
                    }
                }
            }
        },
        "/rides/:id/finish": {
            "post": {
                "description": "finish ride",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rides"
                ],
                "summary": "finishes the ride that matches the given ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ride ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rides.Ride"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.RideRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                },
                "vehicle_id": {
                    "type": "string"
                }
            }
        },
        "rides.Ride": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "finished": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "vehicle_id": {
                    "type": "string"
                }
            }
        }
    }
}